created Entities and inserted values in db.

+++ shop/src/main/resources/db/migration/V1_2__create_product.sql
@@ -2,7 +2,7 @@ CREATE TABLE Product(
 id int primary key auto_increment,
 name varchar(40) not null,
 description varchar(200) not null,
-price numeric not null,
+price double not null,
 weight double not null,
 categoryId int,
 CONSTRAINT FK_ProductCategory foreign key(categoryId) REFERENCES ProductCategory(id),
